Refining the computation process...hopefully
